Debug and Refactor Web App
This commit includes the following changes:

- Separated frontend (HTML/JS) and backend (Python) code.
- Moved Supabase and OpenAI keys to a `.env` file.
- Improved error handling in the JavaScript code.
- Added `python-dotenv` to `requirements.txt`.
- Created a `templates` directory for the HTML file.
- Updated `app.py` to render the HTML template.
